+++ src/main/java/com/borisfarber/instasearch/formats/XmlDecompressor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.borisfarber.instasearch.binary;
+package com.borisfarber.instasearch.formats;
 
 import com.borisfarber.instasearch.controllers.PrivateFolder;
 import com.google.common.io.LittleEndianDataInputStream;
